requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.2
apptools==5.1.0
arcgis==1.9.1
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bokeh==2.4.1
branca==0.4.2
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.1
cairocffi==1.2.0
CairoSVG==2.5.1
Cartopy==0.20.1
certifi==2020.12.5
cffi==1.14.4
cftime==1.4.1
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
configobj==5.0.6
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
cssselect2==0.4.1
cycler==0.10.0
Cython==0.29.23
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
DIRECT==1.0.1
Django==3.2.6
entrypoints==0.3
envisage==6.0.1
et-xmlfile==1.1.0
fastkml==0.11
Fiona==1.8.20
Flask==2.0.1
Flask-Compress==1.10.1
flatbuffers==1.12
folium==0.12.1.post1
fonttools==4.28.5
freegames==2.3.2
future==0.18.2
gast==0.4.0
geographiclib==1.50
geomet==0.3.0
geopandas==0.10.2
geopy==2.2.0
geos==0.2.3
gmplot==1.4.1
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
image==1.5.33
imageio==2.9.0
imgkit==1.2.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
incremental==21.3.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jtutils==0.0.8
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-server==1.13.2
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==21.7.0
kiwisolver==1.3.1
lerc==0.1.0
lesscpy==0.15.0
leven==1.0.4
libpysal==4.6.0
llvmlite==0.36.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
mat73==0.52
matlab==0.1
matplotlib==3.5.1
mayavi==4.7.3
mistune==0.8.4
mpld3==0.5.5
mpmath==1.2.1
multidict==5.2.0
munch==2.5.0
nbclassic==0.3.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netCDF4==1.5.5.1
networkx==2.6.3
nose==1.3.7
notebook==6.2.0
ntlm-auth==1.5.0
numba==0.53.1
numexpr==2.7.3
numpy==1.22.1
oauthlib==3.1.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.1.48
openpyxl==3.0.9
opt-einsum==3.3.0
osm-runner==0.0.3
osmnx==1.1.2
packaging==20.9
Panda3D==1.10.11.post1
pandas==1.3.5
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
plotly-express==0.4.1
plotly-svg==0.11
ply==3.11
pointpats==2.2.0
priority==1.3.0
progress==1.6
prometheus-client==0.9.0
prompt-toolkit==3.0.16
Protego==0.1.16
protobuf==3.15.8
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
PyDispatcher==2.0.5
pyee==8.1.0
pyface==7.3.0
pygame==2.0.1
pygeoif==0.7
Pygments==2.8.0
pygmt==0.5.0
pykml==0.2.0
PyKrige==1.6.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyproj==3.3.0
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.17.3
pyshp==2.1.3
PySimpleGUI==4.46.0
pytemperature==1.0
python-certifi-win32==1.6
python-csv==0.0.13
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
queuelib==1.6.2
rdp==0.8
requests==2.27.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rpy2==3.4.3
rsa==4.7.2
Rtree==0.9.7
scikit-gstat==0.5.4
scikit-learn==0.24.2
scikit-sparse==0.4.5
scipy==1.6.3
Scrapy==2.5.1
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==21.1.0
setuptools-scm==6.3.2
Shapely==1.7.1
simplekml==1.3.6
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
statsmodels==0.12.2
sympy==1.9
TBB==0.2
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc1
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
tinycss2==1.1.0
tomli==2.0.0
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
traits==6.3.1
traitsui==7.2.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
ujson==5.1.0
urllib3==1.26.3
vtk==9.0.3
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==8.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wslink==1.1.0
xarray==0.20.0
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.7.0
zipp==3.6.0
zope.interface==5.4.0
